Add Int8-UInt64 and Float32 numeric dtypes
Eight numeric dtypes join Int64 and Float64 through every operation.
Dispatch loops over NUMERIC_DTYPES at compile time, so each iteration
sees a concrete Scalar[D]: construction, display, equality, sorting,
hashing (group_by, join, unique), casts, CSV, Arrow, and when/then,
fill_null, and comparisons are generic.

Integer arithmetic is checked at each width: widths below 64 bits and
UInt64 compute in 128 bits and range-check, while Int64 keeps its
existing path. Float32 uses the SIMD float kernels. Sums are exact in
128 bits; 8/16-bit sums give Int64 as in Polars, and wider types keep
their type with overflow checks. Reductions keep canonical Int64/Float64
state, with UInt64 re-encoded order-preservingly. Casts go through an
exact Int128/Float64 intermediate; text parses at the target width.
Typed literals: lit(Int32(1)) and generic lit[D](Scalar[D]).

A table-driven test runs every dispatch site for every dtype, plus
per-width overflow, sum, cast, CSV, and Float32 SIMD-width checks.

Closes #30.

+++ tests/test_cast.mojo
@@ -230,7 +230,7 @@ def test_series_cast_masks_and_validation() raises:
         values(series.cast("int64", strict=False)), [String("5"), "null"]
     )
     with assert_raises(contains="Unknown cast dtype"):
-        _ = series.cast("int32")
+        _ = series.cast("int128")
     var frame = DataFrame([series.copy()])
     with assert_raises(contains="Unknown cast dtype: decimal"):
         _ = frame.select(col("s").cast("decimal"))
